Reorganize utility packages into simple-go directory
Move utility packages from internal/ to a top-level simple-go/ directory:
- assert
- logger
- must
- preconditions
- utils

All import paths updated from git.15b.it/eno/critic/internal/... to
git.15b.it/eno/critic/simple-go/...
