+++ internal/mcp/server.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 
 	"git.15b.it/eno/critic/internal/git"
-	"git.15b.it/eno/critic/internal/logger"
+	"git.15b.it/eno/critic/simple-go/logger"
 	"git.15b.it/eno/critic/internal/messagedb"
 	"git.15b.it/eno/critic/pkg/critic"
 )
