+++ internal/messagedb/messaging.go
@@ -3,7 +3,7 @@ package messagedb
 import (
 	"fmt"
 
-	"git.15b.it/eno/critic/internal/logger"
+	"git.15b.it/eno/critic/simple-go/logger"
 	"git.15b.it/eno/critic/pkg/critic"
 )
