+++ simple-go/logger/logger.go
