+++ simple-go/must/must.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"strings"
 
-	"git.15b.it/eno/critic/internal/logger"
+	"git.15b.it/eno/critic/simple-go/logger"
 	"github.com/samber/lo"
 )
