+++ tests/integration/deleted_lines_test.go
@@ -4,9 +4,9 @@ import (
 	"testing"
 
 	"github.com/samber/lo"
-	"git.15b.it/eno/critic/internal/assert"
+	"git.15b.it/eno/critic/simple-go/assert"
 	"git.15b.it/eno/critic/internal/git"
-	"git.15b.it/eno/critic/internal/must"
+	"git.15b.it/eno/critic/simple-go/must"
 	ctypes "git.15b.it/eno/critic/pkg/types"
 )
